requirements.txt
streamlit>=1.36
pdfplumber>=0.11
pymupdf>=1.24
pandas>=2.2
pyyaml>=6.0
